fix(actions/generic): eliminate import cycle
